% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{BUM_dens_sign}
\alias{BUM_dens_sign}
\title{Beta Uniform Mixture density function}
\usage{
BUM_dens_sign(x, lambda = lambda, a = a)
}
\arguments{
\item{x}{x values from the fitted BUM model}

\item{lambda}{lambda from the fitted BUM model}

\item{a}{alpha from the fitted BUM model}
}
\value{
density value of the Beta Uniform Mixture density function
}
\description{
Beta Uniform Mixture density function
}
\examples{
# Internal function
}
