% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{DESeq2_test}
\alias{DESeq2_test}
\title{DESeq2_test: differential expression analysis of RNASeq data using DESeq2}
\usage{
DESeq2_test(
  data,
  column,
  groups,
  reference = "WT",
  covariates = c(),
  speedup = FALSE,
  parallel = FALSE
)
}
\arguments{
\item{data}{data matrix of counts data}

\item{column}{column name of the groups to be used for the definitions}

\item{groups}{design matrix containing the definitions to be compared.}

\item{reference}{Reference group to use in the DE analysis}

\item{covariates}{List of column names in the design matrix to be used as
covariates}

\item{speedup}{Boolean parameter to indicate if the faster version of DESeq2
should be used with fitType = "glmGamPoi" (note: it may not work depending
on the system)}

\item{parallel}{Boolean parameter to indicate if DESeq2 should be run in
parallel mode}
}
\value{
DESeq2 result object with the pvalues, adjusted pvalues and
foldchanges of the DE analysis
}
\description{
DESeq2_test: differential expression analysis of RNASeq data using DESeq2
}
\examples{
# Internal function

}
