% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{dissimilarity_correlation}
\alias{dissimilarity_correlation}
\title{\code{dissimilarity_correlation}}
\usage{
dissimilarity_correlation(
  data1,
  data2,
  outputName = "",
  cond1 = "data1",
  cond2 = "data2",
  perm_dist = NULL,
  perm_dplus = NULL,
  perm_dmin = NULL,
  perm_mode = FALSE
)
}
\arguments{
\item{data1}{dataframe of the foldchange and adjusted pvalue of condition 1
with the columns: fc (foldchange), pval (adjusted pvalue), genes (gene
identifiers)}

\item{data2}{the same as data1 for condition 2}

\item{outputName}{filename for the output visualization}

\item{cond1}{name of condition 1}

\item{cond2}{name of condition 2}

\item{perm_dist}{vector of permutation of the dissimilarity scores}

\item{perm_dplus}{vector of permutation of the dplus values}

\item{perm_dmin}{vector of permutation of the dmin values}

\item{perm_mode}{if TRUE, permutation mode is used to calculate the
permutation distributions. No need to supply the permutation distributions.
No visualizations will be made}
}
\value{
list of the correlation measure (rho), the similarity score (dplus)
and the anti-similarity score (dmin).
}
\description{
\code{dissimilarity_correlation}
}
\examples{
# Internal function

}
