% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{edgeR_test}
\alias{edgeR_test}
\title{edgeR_test: differential expression analysis of RNASeq data using edgeR}
\usage{
edgeR_test(data, gene_names, group, reference)
}
\arguments{
\item{data}{data matrix of counts data}

\item{gene_names}{vector of gene symbols to read to the result dataframe}

\item{group}{column name of the group of interest}

\item{reference}{reference group to use in the DE analysis}
}
\value{
EdgeR result object with the pvalues, adjusted pvalues and
foldchanges of the DE analysis
}
\description{
edgeR_test: differential expression analysis of RNASeq data using edgeR
}
\examples{
# Internal function

}
