% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBRA_method.R
\name{generate_permutation_dist}
\alias{generate_permutation_dist}
\title{Generate CIBRA impact score permutation distribution}
\usage{
generate_permutation_dist(
  data,
  case_list,
  control_list,
  control_definition,
  confidence,
  iterations,
  covariates = c(),
  covariate_matrix = NULL,
  parallel = FALSE,
  method = "DESeq2",
  permutation = "full"
)
}
\arguments{
\item{data}{RNA count dataframe with genes as rows and samples as columns}

\item{case_list}{vector of number of cases to test (vector)}

\item{control_list}{vector of number of controls to test (vector)}

\item{control_definition}{Definition term that will be used as reference for
the comparison (e.g. WT)}

\item{confidence}{alpha threshold to calculate the proportion (default is 0.1)}

\item{iterations}{number of interations (int)}

\item{covariates}{list of column names from the definition matrix to use as covariates (supported only with DESeq2)}

\item{parallel}{boolean status if the run should be done in parallel (boolean)}

\item{method}{method to perform the differential expression analysis, supported methods are DESeq2, edgeR and limma}

\item{permutation}{permutation style to perform, either full where all values
in the matrix are permuted or sample if only the samples should be permuted.}
}
\value{
list of permutation results (dataframe), p-value, adjusted p-value
and foldchange generated from the DE analysis from the permutations
}
\description{
Generate CIBRA impact score permutation distribution
}
\examples{

# load transcriptomics data
count_data <- CIBRA::TCGA_CRC_rna_data

# subset sample to have a quicker run-time for the example
count_data <- count_data[,1:50]

# set parameters for the reference distribution
control_definition <- "NO_SNV"
confidence <- 0.1 # use the same confidence as used in standard analysis
iterations <- 9 # recommended are at least 1000 permutations to explore the full space

# create lists of different case and control sizes
case_list <- seq(10, ncol(count_data), length.out = 3) # recommended value is at least 20 different case and control values
control_list <- seq(10, ncol(count_data), length.out = 3)

# run permutation screen
CIBRA_res <- generate_permutation_dist(data = count_data, case_list = case_list, control_list = control_list, control_definition = control_definition, confidence = confidence, iterations = iterations, parallel = FALSE, permutation = "sample")

}
