% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{Limma_v_test}
\alias{Limma_v_test}
\title{Limma_v_test: differential expression analysis of RNASeq data using Limma + Voom}
\usage{
Limma_v_test(data, column, groups, reference = "WT")
}
\arguments{
\item{data}{data matrix of counts data}

\item{column}{column name of the groups to be used for the definitions}

\item{groups}{design matrix containing the definitions to be compared.}

\item{reference}{reference group to use in the DE analysis}
}
\value{
Limma-Voom result object with the pvalues, adjusted pvalues and
foldchanges of the DE analysis
}
\description{
Limma_v_test: differential expression analysis of RNASeq data using Limma + Voom
}
\examples{
# Internal function

}
