% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIBRA_method.R
\name{permutation_test}
\alias{permutation_test}
\title{CIBRA impact score permutation test}
\usage{
permutation_test(results, perm_dist, test = "perm")
}
\arguments{
\item{results}{observations dataframe returned by CIBRA_run}

\item{perm_dist}{permutation dataframe returned by generate_permutation_dist}

\item{test}{either 'perm' for permutation test or 'gamma' for gamma fitted permutation test}
}
\value{
if test: gamma, list of elements results (dataframe of permutation results),
fit_proportion: gamma fitted function for the proportion and fit_sign_area:
gamma fitted function for the significant area.
Else if test: perm, only a dataframe of the results.
}
\description{
CIBRA impact score permutation test
}
\examples{
# See example run_CIBRA

}
