% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{randomization}
\alias{randomization}
\title{Randomized CIBRA signal measure calculation}
\usage{
randomization(
  data,
  n_cases,
  n_control,
  iterations,
  confidence,
  case,
  control,
  covariates = c(),
  covariate_matrix = NULL,
  parallel = FALSE,
  speedup = FALSE,
  column = "rand",
  method = "DESeq2",
  permutation = "sample"
)
}
\arguments{
\item{data}{RNA count dataframe with genes as rows and samples as columns
(dataframe)}

\item{n_cases}{number of cases in the data (num)}

\item{n_control}{number of controls in the data (num)}

\item{iterations}{number of iterations to run the permutation (num)}

\item{confidence}{confidence (Tau) for the proportion calculation (num)}

\item{case}{case definition (str)}

\item{control}{control definition (str)}

\item{covariates}{list of column names from the definition matrix to use as covariates (supported only with DESeq2)}

\item{covariate_matrix}{design dataframe of the covariates,
columns to take along as covariate values and samples as rownames.}

\item{parallel}{boolean value indicating if the method should be run in
parallel (boolean)}

\item{speedup}{boolean value if the DESeq2 sould be run in speedup mode
(boolean)}

\item{column}{column name to give to the permutated sample column (string)}

\item{method}{DE analysis method to use (options: DESeq2, edgeR and
limma-voom)}

\item{permutation}{permutatin appraoch to use, either sample or full (string)}
}
\value{
list containing 6 variables: the mean random proportion (float), the
standard deviation of the calculated proportions (float), the mean random
significant area (float), the standard deviation of the calculated
significant area (float), signal_data: dataframe of the results containing
the proportion, and significant area for each iteration, pvalue matrix of
the differential expression analysis for each iteration as a column and
genes as rows, adjusted pvalue matrix of the differential expression
analysis for each iteration as a column and genes as rows, foldchange
matrix of the differential expression analysis for each iteration as a
column and genes as rows
}
\description{
Randomized CIBRA signal measure calculation
}
\examples{
# Internal function

}
