% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{scale_distance}
\alias{scale_distance}
\title{\code{scale_distance} calculates the similarity scores
Uses the diffential expression states to calculate the similarity and anti-similarity scores.}
\usage{
scale_distance(data1, data2, genes)
}
\arguments{
\item{data1}{dataframe of the foldchange and adjusted pvalue of condition 1
with the columns: fc (foldchange), pval (adjusted pvalue), genes (gene
identifiers)}

\item{data2}{data1 for condition 2}

\item{genes}{vector of the intersect of genes between data1 and data2}
}
\value{
list containing the similarity scores dplus and dmin and the corresponding genes associated with the dplus and dmin scores, dplusgenes and dmingenes.
}
\description{
\code{scale_distance} calculates the similarity scores
Uses the diffential expression states to calculate the similarity and anti-similarity scores.
}
\examples{
# Internal function
}
