% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{score_assignment}
\alias{score_assignment}
\title{Assign score to the genes depending on in which box they lay in the volcano plot}
\usage{
score_assignment(pval, fc)
}
\arguments{
\item{pval}{vector of adjusted pvalues from DE analysis}

\item{fc}{vector of foldchanges from DE analysis}
}
\value{
vector of score assignments
}
\description{
Assign score to the genes depending on in which box they lay in the volcano plot
}
\examples{
# Internal function
}
