% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{signal_measures}
\alias{signal_measures}
\title{calculate the CIBRA signal measures
Calculate the significant area and proportion from a p-value distribution}
\usage{
signal_measures(
  pvalues,
  confidence,
  gene = NULL,
  cond = NULL,
  outputDir = FALSE
)
}
\arguments{
\item{pvalues}{vector of pvalues generated from DE analysis}

\item{confidence}{confidence value to use to calculate the proportion}

\item{gene}{vector of gene name}

\item{cond}{vector of condition name}

\item{outputDir}{path for directory of result visualizations}
}
\value{
list containing the poportion and significant area
}
\description{
calculate the CIBRA signal measures
Calculate the significant area and proportion from a p-value distribution
}
\examples{
# Internal function

}
