% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_functions.R
\name{volcano_de_state_plot}
\alias{volcano_de_state_plot}
\title{Volcano differential expression state plot}
\usage{
volcano_de_state_plot(
  fc,
  pval,
  link_column,
  condition_column,
  state_column,
  score = NULL,
  pvalue = NULL
)
}
\arguments{
\item{fc}{vector of foldchanges from condition 1 and 2}

\item{pval}{vector of adjusted p-values from condition 1 and 2}

\item{link_column}{gene names}

\item{condition_column}{vector of condition status of the rows}

\item{state_column}{Differential expression states from the
dissimilarity_correlation function}

\item{score}{score to include as annotation for the plot}

\item{pvalue}{pvalue to include as annotation for the plot}
}
\value{
generates a figure
}
\description{
Volcano differential expression state plot
}
\examples{
# Internal function
}
